% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateUnemploymentData.R
\name{updateUnemploymentData}
\alias{updateUnemploymentData}
\title{Get BLS Unemployment Data}
\usage{
updateUnemploymentData(lastYear = 2023)
}
\arguments{
\item{lastYear}{most recent year that data will be pulled from}
}
\description{
This function allows the user to update the unemployment data in the
blsData.sqlite database.
This not only allows the user to pull the latest data but also to update the data.
Since data is typically revised in the months after it is release.
The year including and preceding nine years are pulled from the BLS.
This is because a data set spanning 10 years is the maximum allowed by the
BLS's tier 1 API. Learn more about the API here:
https://www.bls.gov/developers/api_signature.htm#years
The blsAPI package is required and is available from github. See examples.
}
\examples{
\dontrun{
# Install devtools if needed.
install.packages("devtools")

# Install blsAPI package if needed.
library(devtools)
install_github("mikeasilva/blsAPI")

# Pull and update with the most recent BLS unemployment data.
updateUnemploymentData(lastYear = 2023)
}
}
\keyword{API}
\keyword{BLS}
\keyword{Unemployment}
