% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsViewer.R
\name{blsViewer}
\alias{blsViewer}
\title{Run the BLS Viewer}
\usage{
blsViewer()
}
\description{
This function runs the BLS data viewer. Currently this viewer allows
the user to select and view unemployment data.
}
\examples{
\dontrun{
# Run the BLS data viewer.
blsViewer()
}
}
\keyword{BLS}
\keyword{Unemployment}
